[CSL-1131] Move wallet stuff from node/Main.hs to Wallet/Full
